+++ web-common/src/runtime-client/runtime-store.ts
@@ -1,11 +1,8 @@
 import { get, writable } from "svelte/store";
-import { goto } from "$app/navigation";
 import {
-  runtimeServiceGetInstance,
-  runtimeServiceUnpackEmpty,
-} from "./gen/runtime-service/runtime-service";
-import { EMPTY_PROJECT_TITLE } from "../features/welcome/constants";
-import { isProjectInitialized } from "@rilldata/web-common/features/welcome/is-project-initialized";
+  isProjectInitialized,
+  handleUninitializedProject,
+} from "@rilldata/web-common/features/welcome/is-project-initialized";
 
 export interface JWT {
   token: string;
@@ -32,9 +29,8 @@ export const projectInitialized = (() => {
     if (!initialized) {
       await handleUninitializedProject();
     } else if (window.location.pathname === "/welcome") {
-      await goto("/");
+      window.location.replace("/");
     }
-
     set(initialized);
   };
 
@@ -48,26 +44,3 @@ export const projectInitialized = (() => {
     },
   };
 })();
-
-async function handleUninitializedProject() {
-  const instanceId = get(runtime).instanceId;
-  // If the project is not initialized, determine what page to route to dependent on the OLAP connector
-  const instance = await runtimeServiceGetInstance(instanceId, {
-    sensitive: true,
-  });
-  const olapConnector = instance.instance?.olapConnector;
-  if (!olapConnector) {
-    throw new Error("OLAP connector is not defined");
-  }
-
-  // DuckDB-backed projects should head to the Welcome page for user-guided initialization
-  if (olapConnector === "duckdb") {
-    await goto("/welcome");
-  } else {
-    // Clickhouse and Druid-backed projects should be initialized immediately
-    await runtimeServiceUnpackEmpty(instanceId, {
-      title: EMPTY_PROJECT_TITLE,
-      force: true,
-    });
-  }
-}
